Update my configuration with Tap Dance actions.
The most important change is the double action home_end key.

One nice idea that might come from it would be to make a volume key
that gets progressively more aggressive on re-tap.

For example, send 2^n times the key on each tap, with n = 1
incrementing on each tap in the same direction, and decrementing on
all changes. It could also work with a less aggressive curve.